requirements.txt
paddlepaddle>=2.5.0
paddleocr>=2.7.0
PyQt6>=6.4.0
Pillow>=9.0.0
pymupdf>=1.23.0
